All the User autentication is up to date with tokens JWT

+++ resources/item.py
@@ -6,6 +6,7 @@ from flask_smorest import Blueprint, abort
 
 from schemas import ItemSchema, ItemUpdateSchema
 
+from flask_jwt_extended import jwt_required,get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 from models import ItemModel
 from db import db
@@ -15,12 +16,19 @@ blp = Blueprint("Items", "items", description="Operations on items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        # We need to be and admin to access this endpointru
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -43,10 +51,12 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
